add: displayImage component and add with offer
