+++ admin/src/pages/offers/CreateOfferModal.js
@@ -2,28 +2,29 @@ import React, { useEffect, useState } from "react";
 import { Modal, Form, Input, message, Upload, Button, Image } from "antd";
 import { UploadOutlined, CloseCircleOutlined } from "@ant-design/icons";
 import { createOffer, deleteOfferImage } from "../../client/offers.client";
+import DisplayImage from "../../components/DisplayImage";
 
-const imageStyle = {
-  display: "inline-block",
-  position: "relative",
-};
+// const imageStyle = {
+//   display: "inline-block",
+//   position: "relative",
+// };
 
-const titleStyle = {
-  display: "inline-block",
-  position: "absolute",
-  top: "40%",
-  width: "300px",
-  margin: "0px 20px",
-};
+// const titleStyle = {
+//   display: "inline-block",
+//   position: "absolute",
+//   top: "40%",
+//   width: "300px",
+//   margin: "0px 20px",
+// };
 
-const deleteButtonStyle = {
-  cursor: "pointer",
-  position: "absolute",
-  marginLeft: "325px",
-  top: "40%",
-  fontSize: "25px",
-  color: "red",
-};
+// const deleteButtonStyle = {
+//   cursor: "pointer",
+//   position: "absolute",
+//   marginLeft: "325px",
+//   top: "40%",
+//   fontSize: "25px",
+//   color: "red",
+// };
 
 export default function CreateOfferModal({ visible, onCreate, onCancel }) {
   const [form] = Form.useForm();
@@ -176,7 +177,7 @@ export default function CreateOfferModal({ visible, onCreate, onCancel }) {
             <Input.TextArea />
           </Form.Item>
 
-          {offerImages &&
+          {/* {offerImages &&
             offerImages.map((offerImage) => (
               <div key={offerImage.id}>
                 <div style={imageStyle}>
@@ -197,7 +198,15 @@ export default function CreateOfferModal({ visible, onCreate, onCancel }) {
                   />
                 </div>
               </div>
-            ))}
+            ))} */}
+
+          <DisplayImage
+            imageArray={offerImages}
+            token={token}
+            deleteImage={deleteOfferImage}
+            handleImageFromState={handleImageFromState}
+          />
+
           <Upload
             name="offerImages"
             action="http://localhost:4000/api/v1/offer-images"
